next

+++ server/__main__.py
@@ -313,12 +313,12 @@ def compute_frame_indexes(
         frame_skip = max(1, round(fps / b.fps))
         final_fps = round(fps / frame_skip)
 
-    start_frame = a.clip_start[0] * fps + a.clip_start[1]
-    end_frame = start_frame + a.clip_duration[0] * fps + a.clip_duration[1]
+    start_frame = int(a.clip_start[0] * fps + a.clip_start[1])
+    end_frame = int(start_frame + a.clip_duration[0] * fps + a.clip_duration[1])
     frame_indexes = []
     i = 0
     while 1:
-        frame_index = start_frame + i * frame_skip
+        frame_index = int(start_frame + i * frame_skip)
         if frame_index > frame_count - 1:
             break
         if frame_index > end_frame - 1:
@@ -397,7 +397,7 @@ def group_by(a: list, size: int):
 
 def parse_time(a: str) -> tuple[int, int]:
     h, m, s, rest = map(int, a.split(":"))
-    return h * 60 * 60 + m * 60 + s, rest
+    return int(h * 60 * 60 + m * 60 + s), rest
 
 
 __main__()
